Implement support for tuple compression perf test
Implement tuple compression perf test in enterprise version
of tarantool. This commit add necessary changes in opensource
version.

NO_DOC=performance test
NO_CHANGELOG=performance test
NO_TEST=performance test
