+++ perf/CMakeLists.txt
@@ -14,3 +14,9 @@ include_directories(${EXTRA_CORE_INCLUDE_DIRS})
 add_executable(tuple.perftest tuple.cc
                ${PROJECT_SOURCE_DIR}/test/unit/box_test_utils.c)
 target_link_libraries(tuple.perftest core box tuple benchmark::benchmark)
+
+foreach(PERF_TEST_SUITE ${EXTRA_PERF_TEST_SUITES})
+    add_subdirectory(
+        "${EXTRA_PERF_TEST_SUITES_SOURCE_DIR}/${PERF_TEST_SUITE}"
+        "${EXTRA_PERF_TEST_SUITES_BINARY_DIR}/${PERF_TEST_SUITE}")
+endforeach()
