+++ perf/tuple.cc
@@ -7,6 +7,8 @@
 #include <iostream>
 #include <benchmark/benchmark.h>
 
+#include "common.h"
+
 const size_t NUM_TEST_TUPLES = 4096;
 const size_t MAX_TUPLE_DATA_SIZE = 512;
 
@@ -431,25 +433,3 @@ tuple_tuple_compare_hint(benchmark::State& state)
 BENCHMARK(tuple_tuple_compare_hint);
 
 BENCHMARK_MAIN();
-
-static void
-show_warning_if_debug()
-{
-#ifndef NDEBUG
-	std::cerr << "#######################################################\n"
-		  << "#######################################################\n"
-		  << "#######################################################\n"
-		  << "###                                                 ###\n"
-		  << "###                    WARNING!                     ###\n"
-		  << "###   The performance test is run in debug build!   ###\n"
-		  << "###   Test results are definitely inappropriate!    ###\n"
-		  << "###                                                 ###\n"
-		  << "#######################################################\n"
-		  << "#######################################################\n"
-		  << "#######################################################\n";
-#endif // #ifndef NDEBUG
-}
-
-struct DebugWarning {
-	DebugWarning() { show_warning_if_debug(); }
-} debug_warning;
